cherry/kokotown/src/com/internousdev/kokotown/dto/PurchaseHistoryInfoDTO.java
package com.internousdev.kokotown.dto;

public class PurchaseHistoryInfoDTO {

	public String productName;

	public String productNameKana;

	public String imageFilePath;

	public String price;

	public String releaseCompany;

	public String releaseDate;

	public String priceComma;

	public int productCount;

	public int id;


	public String getProductName() {
		return productName;
	}

	public void setProductName(String productName) {
		this.productName = productName;
	}

	public String getProductNameKana() {
		return productNameKana;
	}

	public void setProductNameKana(String productNameKana) {
		this.productNameKana = productNameKana;
	}

	public String getImageFilePath() {
		return imageFilePath;
	}

	public void setImageFilePath(String imageFilePath) {
		this.imageFilePath = imageFilePath;
	}

	public String getPrice() {
		return price;
	}

	public void setPrice(String price) {
		this.price = price;
	}

	public String getReleaseCompany() {
		return releaseCompany;
	}

	public void setReleaseCompany(String releaseCompany) {
		this.releaseCompany = releaseCompany;
	}

	public String getReleaseDate() {
		return releaseDate;
	}

	public void setReleaseDate(String releaseDate) {
		this.releaseDate = releaseDate;
	}

	public String getPriceComma() {
		return priceComma;
	}

	public void setPriceComma(String priceComma) {
		this.priceComma = priceComma;
	}

	public int getProductCount() {
		return productCount;
	}

	public void setProductCount(int productCount) {
		this.productCount = productCount;
	}

	public int getId() {
		return id;
	}

	public void setId(int id) {
		this.id = id;
	}



}
